refactor: create module for credentials/config file i/o
This creates a new module that focuses on the i/o of
the credentials/config file. Mainly the reading and
writing of data to the file.

- Also moves the get credentials/config file name functions
  in to it
