+++ src/credentials/providers/sharedCredentialsProviderFactory.ts
@@ -5,15 +5,11 @@
 
 import * as fs from 'fs-extra'
 import { getLogger, Logger } from '../../shared/logger'
-import {
-    getConfigFilename,
-    getCredentialsFilename,
-    loadSharedCredentialsSections,
-    updateAwsSdkLoadConfigEnvVar,
-} from '../sharedCredentials'
+import { loadSharedCredentialsSections, updateAwsSdkLoadConfigEnvVar } from '../sharedCredentials'
 import { CredentialsProviderType } from './credentials'
 import { BaseCredentialsProviderFactory } from './credentialsProviderFactory'
 import { SharedCredentialsProvider } from './sharedCredentialsProvider'
+import { getCredentialsFilename, getConfigFilename } from '../sharedCredentialsFile'
 
 export class SharedCredentialsProviderFactory extends BaseCredentialsProviderFactory<SharedCredentialsProvider> {
     private readonly logger: Logger = getLogger()
