+++ src/credentials/sharedCredentials.ts
@@ -4,23 +4,10 @@
  */
 
 import * as vscode from 'vscode'
-import { join } from 'path'
-import { EnvironmentVariables } from '../shared/environmentVariables'
 import { SystemUtilities } from '../shared/systemUtilities'
 import { ToolkitError } from '../shared/errors'
 import { assertHasProps } from '../shared/utilities/tsUtils'
-
-export function getCredentialsFilename(): string {
-    const env = process.env as EnvironmentVariables
-
-    return env.AWS_SHARED_CREDENTIALS_FILE || join(SystemUtilities.getHomeDirectory(), '.aws', 'credentials')
-}
-
-export function getConfigFilename(): string {
-    const env = process.env as EnvironmentVariables
-
-    return env.AWS_CONFIG_FILE || join(SystemUtilities.getHomeDirectory(), '.aws', 'config')
-}
+import { getConfigFilename, getCredentialsFilename } from './sharedCredentialsFile'
 
 export async function updateAwsSdkLoadConfigEnvVar(): Promise<void> {
     const configFileExists = await SystemUtilities.fileExists(getConfigFilename())
