+++ src/test/credentials/sharedCredentials.test.ts
@@ -6,9 +6,9 @@
 import * as assert from 'assert'
 import * as path from 'path'
 import * as fs from 'fs-extra'
-import { getConfigFilename, getCredentialsFilename } from '../../credentials/sharedCredentials'
 import { EnvironmentVariables } from '../../shared/environmentVariables'
 import { makeTemporaryToolkitFolder } from '../../shared/filesystemUtilities'
+import { getCredentialsFilename, getConfigFilename } from '../../credentials/sharedCredentialsFile'
 
 describe('sharedCredentials', function () {
     let tempFolder: string
